# contracts/openapi.yaml
components:
  schemas:
...
    CreateFiscalYearRequest:
      description: Request body for creating a fiscal year.
      properties:
        end_date:
          description: End date (YYYY-MM-DD).
          example: 2026-12-31
          format: date
          type: string
        name:
          description: Fiscal year name (e.g., "FY 2026").
          example: FY 2026
          type: string
        start_date:
          description: Start date (YYYY-MM-DD).
          example: 2026-01-01
          format: date
          type: string
      required:
      - name
      - start_date
      - end_date
      type: object
...
    FiscalPeriodResponse:
      description: Response for a fiscal period.
      properties:
        end_date:
          description: End date.
          format: date
          type: string
        id:
          description: Period ID.
          format: uuid
          type: string
        is_adjustment_period:
          description: Whether this is an adjustment period.
          type: boolean
        name:
          description: Period name.
          example: January 2026
          type: string
        period_number:
          description: Period number within the fiscal year.
          example: 1
          format: int32
          type: integer
        start_date:
          description: Start date.
          format: date
          type: string
        status:
          description: 'Status: open, soft_close, or closed.'
          example: open
          type: string
      required:
      - id
      - name
      - period_number
      - start_date
      - end_date
      - status
      - is_adjustment_period
      type: object
    FiscalYearResponse:
      description: Response for a fiscal year with periods.
      properties:
        end_date:
          description: End date.
          format: date
          type: string
        id:
          description: Fiscal year ID.
          format: uuid
          type: string
        name:
          description: Fiscal year name.
          example: FY 2026
          type: string
        periods:
          description: Nested periods.
          items:
            $ref: '#/components/schemas/FiscalPeriodResponse'
          type: array
        start_date:
          description: Start date.
          format: date
          type: string
        status:
          description: 'Status: open or closed.'
          example: open
          type: string
      required:
      - id
      - name
      - start_date
      - end_date
      - status
      - periods
      type: object
...
    UpdatePeriodStatusRequest:
      description: Request body for updating period status.
      properties:
        status:
          description: 'New status: "open", "soft_close", or "closed".'
          example: closed
          type: string
      required:
      - status
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: HTTP API layer with Axum routes and middleware
  license:
    name: ''
  title: zeltra-api
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /organizations/{org_id}/fiscal-periods/{period_id}/status:
    patch:
      operationId: update_period_status
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        schema:
          format: uuid
          type: string
      - description: Fiscal Period ID
        in: path
        name: period_id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePeriodStatusRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiscalPeriodResponse'
          description: Period status updated successfully
        '400':
          description: Invalid status or transition
        '403':
          description: Forbidden
        '409':
          description: Earlier periods must be closed first
      security:
      - bearerAuth: []
      summary: PATCH `/organizations/{org_id}/fiscal-periods/{period_id}/status` - Update period status.
      tags:
      - Fiscal
  /organizations/{org_id}/fiscal-years:
    get:
      operationId: list_fiscal_years
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FiscalYearResponse'
                type: array
          description: List of fiscal years
        '403':
          description: Forbidden
      security:
      - bearerAuth: []
      summary: GET `/organizations/{org_id}/fiscal-years` - List fiscal years with nested periods.
      tags:
      - Fiscal
    post:
      operationId: create_fiscal_year
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFiscalYearRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiscalYearResponse'
          description: Fiscal year created successfully
        '400':
          description: Invalid date range
        '403':
          description: Forbidden
        '409':
          description: Overlapping fiscal year
      security:
      - bearerAuth: []
      summary: POST `/organizations/{org_id}/fiscal-years` - Create a fiscal year with auto-generated periods.
      tags:
      - Fiscal
...